requirements.txt
python-telegram-bot==21.6
sympy==1.13.1
